requirements.txt
Django>=5.0.6
djangorestframework>=3.15.1
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.27.2
python-dotenv>=1.0.1
pillow>=10.3.0
PyYAML>=6.0.1
gunicorn==22.0.0

